Layout Manager
This app starts in the console, and I'd like to try to approach it in a much organised way, that's why we'll start by creating a Layout Manager, and only afterwards we're going to focus on features like the text editor.
The app at the end is planned to have a proper GUI, but it starts in the console.