Adding Text Editor logic
The LayoutMod reached a mid-maxed state, where it could be used to draw somethings, but it will be developed further.
The TextMod has been created, aiming to implement cursor movement and text edition logic. It's now at it's initial phase, meaning that for it to become reliable more work is needed