char[] to store text
I made some very important decisions about how this app will be developed, and I wanted to share those decisions: First - I intend to focus more on testing during this project. I'd like also to use char[] to store the editor content, which gives the lowest-level experience, so that's the plan, but I reserve the right for it to change.
I also added the ability to do spaces between words in the editor.