Pushing text to editor
The implementation of the pipeline that pushes text changes into the console's element, started today.
UpdateContent() was added, which aims to format the text so it can be properly displayed in the console editor window. Also modified the AffirmFixedContent(), in a way that would make it more compatible with the applied changes. Yet, the text is not displayed in the console window.