General Leap Forward
I started building a more integrated pipeline to draw the content on the screen, by:
Creating methods that abstract (while keeping tight control and efficiency) copying stuff inside char[], regardless if it's somewhere in the middle of it, or added to the end.
Expanded TextPointer, to keep track of more variables that might be needed to track the position of the cursor, and enable scrolling in the future (if Lines > than Element's Height).
Added static class ArrayTooling, to tie Array manipulation within a single class.

Despite the changes, the app is still non-functional, this might change soon enough.